Updated last minute changes, added icon, name, changed logo background and fixed typedefs
