+++ client/public/index.html
@@ -2,7 +2,7 @@
 <html lang="en">
   <head>
     <meta charset="utf-8" />
-    <link rel="icon" href="%PUBLIC_URL%/favicon.ico" />
+    <link rel="icon" href="%PUBLIC_URL%/smallLogo.png" />
     <link rel="preconnect" href="https://fonts.googleapis.com">
 <link rel="preconnect" href="https://fonts.googleapis.com">
 <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
@@ -19,9 +19,9 @@
     <meta name="theme-color" content="#000000" />
     <meta
       name="description"
-      content="Web site created using create-react-app"
+      content="Lai Saetern is an independent hairstylist at Tangelo Hair Salon. They specialize in cut, color and styling for women and men."
     />
-    <title>ENTER COMPANY NAME HERE</title>
+    <title>Lai Hairstylist</title>
   </head>
   <body>
     <div id="root"></div>
