+++ client/src/components/book/BookContainer.js
@@ -5,7 +5,7 @@ import { Col, Row } from 'reactstrap';
 import '../../css/booking/booking.css';
 
 // import BookingForm from './BookingForm.js';
-import Calendar from './Calendar2.0.js';
+import Calendar from './Calendar.js';
 import Hours from "../about/Hours.js";
 
 function Book() {
