+++ client/src/components/book/Calendar.js
@@ -1,78 +1,89 @@
 import React from 'react';
+import {useEffect, useState} from 'react'
 import FullCalendar from '@fullcalendar/react'; // must go before plugins
 import listPlugin from '@fullcalendar/list'; // a plugin!
-import { useQuery } from '@apollo/client';
+import { useQuery, useMutation } from '@apollo/client';
 import { Button, Modal, ModalHeader, ModalBody, ModalFooter } from "reactstrap";
 import {QUERY_AVAILABLE_APPT} from "../../utils/queries"
-const {loading, error, data} = useQuery(QUERY_AVAILABLE_APPT)
+import {BOOK_APPT} from "../../utils/mutations"
 
 
-class Calendar extends React.Component {
-  constructor(props){
-    super(props);
-    this.state = {
-      modal: false,
-      calendarWeekends: true,
-      event: []
-    };
-    this.handleEventClick = this.handleEventClick.bind(this)
-    this.handleButton = this.handleButton.bind(this)
-  }
-  bookAppt(id){
+function Calendar(props) {
+    const [modal,setModal] = useState(false)
+    const [calendarWeekends,setCalendarweekends] = useState(true)
+    const [event, updateEvent] = useState([])
+    const {loading, error, data} = useQuery(QUERY_AVAILABLE_APPT,{
+      variables: {
+        avail: true,
+      },
+      fetchPolicy: "network-only"
+    })
+    const [bookAppt, {reset}] = useMutation(BOOK_APPT)
     
-  }
-  toggle = () => {
-    this.setState({ modal: !this.state.modal });
-  };
+    let initialEvents = data?.appts || [];
+
+    const activateModal = () => {
+     setModal(true);
+    }
+    const closeModal = () => {
+        setModal(false)
+    }
   
-  handleButton = () => {
-    this.bookAppt(this.state.event.url)
-  };
-  handleEventClick = (info) => {
-    console.log(info.event)
-    let id = info.event.id
-    let url = '/api/appointment/' + id
-    let title = info.event.title
-    let date = info.event.start.toDateString()
-    let time = info.event.start.toLocaleTimeString()
-    this.toggle();
-    this.setState({event: {url, id, date, title, time}})
-    console.log(this.state)
-  };
-  render(){
+   
+  const handleButton = () => {
+     console.log(event)
+     bookAppt({variables: { id: event.id}})
+     closeModal()
+     window.location.reload()
+   };
+
+     const handleEventClick = (info) => {
+     let id = info.event.extendedProps._id
+     let title = info.event.title
+     let date = info.event.start.toDateString()
+     let time = info.event.start.toLocaleTimeString()
+     updateEvent({id, date, title, time})
+     activateModal()
+   };
     return (
       <div>
-      <FullCalendar
-        plugins={[ listPlugin ]}
-        initialView="listWeek"
-        noEventsContent="No available appointments at this time, please check back soon!"
-        eventClick={this.handleEventClick}
-        // events= {...apptArray}
-        />
+          {loading ? (
+            <div>Loading...</div>
+          ) : error ? (
+            <div>We're sorry, online appointment booking is not available at this time, please call <a href='tel:804-222-1111' >(Not) are-alnum</a> to setup your appt</div>
+          ) : (
+              <FullCalendar
+                plugins={[ listPlugin ]}
+                initialView="listWeek"
+                noEventsContent="No available appointments at this time, please check back soon!"
+                eventClick={handleEventClick}
+                rerenderEvents
+                initialEvents = {initialEvents}
+                />
+          )}
         <Modal
-              isOpen={this.state.modal}
-              toggle={this.toggle}
+              isOpen = {modal}
+              onClosed = {() => reset()}
             >
-              <ModalHeader toggle={this.toggle}>
-                {this.state.event.title}
+              <ModalHeader closeModal>
+                {event.title}
               </ModalHeader>
               <ModalBody>
                 <div>
-                    {this.state.event.date}
+                    {event.date}
                 </div>
               <div>
-                    {this.state.event.time}
+                    {event.time}
                 </div>
               </ModalBody>
               <ModalFooter>
-                <Button color="primary" onClick={this.handleButton}>Claim Appt</Button>
-                <Button color="secondary" onClick={this.toggle}>
+                <Button color="primary" onClick={handleButton}>Claim Appt</Button>
+                <Button color="secondary" onClick={closeModal}>
                   Cancel
                 </Button>
               </ModalFooter>
             </Modal>
   </div>
     )
-  }
 }
 export default Calendar
